use request instead of xmlhttprequest for node
This gets a few steps closer -- can now actually receive all the login
data when running login() in node.  It looks hella ugly, but it works.

Have not tested this code revision in browser yet.

For some reason, despite request normally wanting to follow redirects,
it does not follow them all, and therefore, we have to add in another
request to get to the right place, when running in node.

Remaining steps:
- Test this in browser.
- Consider using 'browser-request' so we hopefully reduce the amount of
split code, just to which module to select, and which login to run
(as i don't expect browser-request is going to have the same problems
that running outside of browser does)
- Figure out some way to get the 'eval' in login.js to work correctly in
Node. Browser eval is currently erroring on my account, with a syntax
error, and no help in identifying exactly what it is.  The code that
it's attempting to eval is messy, but works in browser. wtf?
- May have to replace it with the old SynerGV 1 method that was quite
easily breakable, of turning the string being evaled into a string that
can be JSON.parsed instead.