code cleanup, de-duplicate some bits, etc
Add getElementsShims -- contains functions getElementsByClassName,
getElementsByName, getCDATASectionsByTagName

getElementsByClassName is a shim to add support for the standard use of
getElementsByClassName to xmldom, getElementsByName is same for the same
named standard HTML dom function.

getCDATASectionsByTagName returns an Array or Array-like containing all
Elements in the document by the given tag... and it looks like I intended
to make it return the entire CDATA section, but it actually currently
returns everything between an open brace { and a ]]> .. so it's actually
return just a JSON-looking section. Oops. Looks like I'll need to update
that function some.

getRequest, putRequest, and getBox now import and use the functions in
getElementsShims instead of having their own copies.