made a bigger mess out of things while cleaning and learning :| heh
Went to create a method to have the library deal with tokens, because
I was pretty well sick of having to pass them in everywhere.  Figured
while I was doing that, since I was trying to make the second to last
parameter to all of the functions optional.. that I would also attempt
to start converting to using Promises instead of old-school callbacks.

That actually all went fine, but I discovered that the change previous
to this actually didn't work at all, and that my test method was
completely not correct -- apparently, when I was testing the previous
change, i was actually testing the code prior to that, because i think
I forgot to rebuild the library each time.  So it looked like it worked,
but it really didn't.  And this is why the ability to test the code from
node is so important.  And when we do test it, we need to make sure that
we are actually testing the current version that the tests are running
inside, since there is this silly build process required, because node
still isn't 100% compliant with all the functions we want/need to use.

SO, lots of lessons learned.

samples/login and samples/getbox are now Promise based.
getBox API no longer takes a callback, but returns a Promise instead.
Other APIs will be switched over as time permits.
Plenty of errors in getElementsShims were discovered, and fixed

Added a tokenStore module.  If you call tokenStore.setTokens() with
the results from the login() API, then you don't need to call getBox
(and eventually all the rest of the API) with a tokens arg.  You will
still be ABLE to use a tokens arg, in case your application wants to be
able to handle requests for multiple accounts or some such thing.

That is just a simple in-memory store. It does not save the given tokens
to anywhere persistant, nor does it currently make any attempts to validate
their legitimacy or anything like that.

So, now I've learned how to deal with errors in Promises, and a whole
bunch of other fun stuff today. Yay!

Oh, Happy New Year everyone!