minor improvements to testsuite
Check that test conversation has 2 messages, and warn if anything other
than expected.

Output all the data from the test conversation when it is first retrieved,
including the contents of messages

call DeleteForever on the test message at the end of testing, so completed
runs should be all clean at the end.