testsuite: add starMessage
fix starMessage to use Promises / tokenStore
fix starMessage to use correct value for star parameter

bump module version to 0.1.0 as this should make all of the "normal"
functionality complete, at least in node -- need to re-write the test
browser app to use all the API changes that occurred during the process
of writing the command line testsuite.

Modules that remain not in testsuite:

checkContacts - requires getting the "xpc" token to work in the node.js
path for login

checkMessages - also requires xpc token

contactQuickAdd - this may end up being removed, it is a holdover from
SynerGV 1.0, which was written before the Google Contacts API change
several years ago, so I'm not sure if this operation is supported or not

donate - need to examine message data to figure out what changes when
a message is donated, as currently I don't know what properties to test
to see if a voicemail has been donated or not

editDefaultForwarding - still need to investigate what this call even does
forward - may not be testable beyond ensuring that an "ok" result is
received, as this function dumps a message to an email account, and I
think that adding email checking functionality to this test suite is WAY
out of scope.

generalSettings - also still need to investigate what this call does

getContacts - requires xpc token

getDoNotDisturb - this should probably be next on the list of things,
it probably just needs to be ported to the current API methods

getTranscriptTiming - may get removed -- not sure if this function has
ever worked? is it supported at the server side still?

searchMessages - this probably needs significant implementation details
still.  It may also be exactly the same as getBox just with a query
string and a different URI.  Needs investigation.

vmDownload - still needs investigation, do not know how/if this call works