v0.1.2: fix check*, getContacts, add lots of notes to sendMessage
tested checkContacts, checkMessages, getContacts, all of which require
the xpc token, which the node login does not currently retrieve.. via
browser.  These functions work.  Just not in node, just yet, because the
login process is not precisely the same in node vs browser.

Found a ton of useful information in the sendMessage command, which
I don't think anyone else has ever attempted to document. So, I attempted
to document it.

Also added a note in login about running the lib as if you're a mobile
browser.  Don't do it, for now.