checkMessages uses voice api not xpc api
checkMessages uses the equivalent GV call not the XPC call, so it works
in Node now.

Added checkMessages to testsuite

Added 200ms delays in many places throughout the test suite, as on my
test account, a bunch of operations started failing when fired in rapid
succession.

Did a bunch of snooping around to API calls that I know of, but don't
know what they do. Documented finds in uris.js